requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pydantic
google-auth
